Small grass plugin cleanup

+++ plugins/grass/worker.js
@@ -49,7 +49,7 @@ const _isPointInTriangle = (p, tri) => {
   const sign = A < 0 ? -1 : 1;
   const s = (p0.y * p2.x - p0.x * p2.y + (p2.y - p0.y) * p.x + (p0.x - p2.x) * p.y) * sign;
   const t = (p0.x * p1.y - p0.y * p1.x + (p0.y - p1.y) * p.x + (p1.x - p0.x) * p.y) * sign;
-  
+
   return s > 0 && t > 0 && (s + t) < 2 * A * sign;
 };
 const _isPointInTriangles = (p, ts) => {
@@ -119,7 +119,7 @@ const grassTextureAtlas = (() => {
             _isPointInTriangles(
               new THREE.Vector2(dx / TEXTURE_CHUNK_SIZE, 1 - (dy / TEXTURE_CHUNK_SIZE)),
               triangles
-            ) ? Float32Array.from(
+            ) ? Uint8Array.from(
               baseColor.clone()
                 .multiplyScalar(0.3 + ((1 - (dy / TEXTURE_CHUNK_SIZE)) * 1))
               .toArray()
